mtstate: collect finger movements until explicit clean
Our touch panel emits finger movements _per finger_.
This means that a regular scroll is emitted in 2 blocks:
1st finger moves.
2nd fingers comes by.

With the traditional uncached detection, this move results
in 2 oppositte rotations, which should never reach the configured
limits.

With finger movement collecting, the movements are accumulated
until one gesture reaches it's configured limit.
When a button is sent, the accumulated movement is cleared.